+++ algorithms/sudoku/with_numpy.py
